refactor: move page data to separate files
