+++ acm_website/src/data/leadership.ts
@@ -1,0 +1,80 @@
+interface PersonData {
+  name: string;
+  role: string;
+  bio: string;
+  linkedin: string;
+  imagePath?: string;
+}
+
+export const leadershipData: PersonData[] = [
+  {
+    name: "Spencer Ye",
+    role: "President",
+    bio: "2027",
+    linkedin: "https://www.linkedin.com/in/spencerye/",
+    imagePath: "/images/team/spencer.jpg"
+  },
+  {
+    name: "Elizabeth Fu",
+    role: "Vice President",
+    bio: "2027",
+    linkedin: "https://www.linkedin.com/in/elizabeth-f-773b78244/",
+    imagePath: "/images/team/elizabeth.jpg"
+  },
+  {
+    name: "Ria Talwar",
+    role: "Treasurer",
+    bio: "2027",
+    linkedin: "https://www.linkedin.com/in/ria-talwar-cs2027/",
+    imagePath: "/images/team/ria.jpg"
+  },
+  {
+    name: "Mischa Kumar",
+    role: "Secretary",
+    bio: "2028",
+    linkedin: "https://www.linkedin.com/in/mischa-kumar-274631298/",
+    imagePath: "/images/team/mischa.jpg"
+  },
+  {
+    name: "Timothy Lin",
+    role: "ICPC Head",
+    bio: "2026",
+    linkedin: "https://www.linkedin.com/in/tlin2004/",
+    imagePath: "/images/team/timothy.jpg"
+  },
+  {
+    name: "Liam Finn",
+    role: "Outreach Chair",
+    bio: "2027",
+    linkedin: "https://www.linkedin.com/in/liam-finn-11o37/",
+    imagePath: "/images/team/liam.jpg"
+  },
+  {
+    name: "Tianai Yue",
+    role: "Events Chair",
+    bio: "2026",
+    linkedin: "https://www.linkedin.com/in/tianai-yue-70981523b/",
+    imagePath: "/images/team/tianai.jpg"
+  },
+  {
+    name: "Alan Li",
+    role: "General Officer",
+    bio: "2026",
+    linkedin: "https://www.linkedin.com/in/alanli6/",
+    imagePath: "/images/team/alan.jpg"
+  },
+  {
+    name: "Vrinda Sehgal",
+    role: "General Officer",
+    bio: "2028",
+    linkedin: "https://www.linkedin.com/in/vrinda-sehgal-0000000000/",
+    imagePath: "/images/team/vrinda.jpg"
+  },
+  {
+    name: "Vicki Chen",
+    role: "General Officer",
+    bio: "2028",
+    linkedin: "https://www.linkedin.com/in/vickichenn/",
+    imagePath: "/images/team/vicki.jpg"
+  }
+];
